Removed unused JarResourceHandler and renamed the new one
